// src/shared/types.ts
import { ReactElement } from "react";

export enum SelectedPage {
  Home = "home",
  Benefits = "benefits",
  OurClasses = "ourclasses",
  ContactUs = "contactus",
}

export interface BenefitType {
  icon: ReactElement,
  title: string,
  description: string
};

export interface ClassType {
  name: string;
  description?: string;
  image: string;
};